[ET-VK] Separate `PhysicalDevice`/`DeviceHandle` from `Adapter.*`
`Adapter.*` has too much going on.

Differential Revision: [D59540131](https://our.internmc.facebook.com/intern/diff/D59540131/)

[ghstack-poisoned]
